requirements.txt
asgiref==3.11.0
attrs==26.1.0
black==26.3.1
click==8.3.2
colorama==0.4.6
Django==6.0.1
django-debug-toolbar==6.3.0
django-rest-framework==0.1.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyboard==0.13.5
MouseInfo==0.1.3
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
psycopg==3.3.3
psycopg-binary==3.3.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.12.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.2
pytokens==0.4.1
pytweening==1.2.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
yt-dlp==2025.11.12
